requirements.txt
appdirs==1.4.3
#apturl==0.5.2
asgiref==3.3.4
astroid==2.6.2
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
#Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
#chrome-gnome-shell==0.0.0
#cliapp==1.20180812.1
Click==7.0
#cmdtest==0.32+git
colorama==0.4.3
#command-not-found==0.3
cryptography==2.8
#cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.3.0
decorator==5.0.9
#defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
#distro-info===0.23ubuntu1
Django==3.2.3
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-taggit==1.5.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
#duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.2
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
keyring==18.0.1
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
#louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.0
matplotlib-inline==0.1.2
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mysqlclient==1.4.4
netifaces==0.10.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.7
packaging==20.3
paramiko==2.6.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.19
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.9.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyRFC3339==1.1
#python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
#python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
ssh-import-id==5.10
supervisor==4.1.0
systemd-python==234
tablib==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
#ttystatus==0.38
#ubuntu-advantage-tools==27.2
#ubuntu-drivers-common==0.0.0
#ufw==0.36
#unattended-upgrades==0.1
urllib3==1.25.8
#usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
#xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
